Remove old deploy functionality
