+++ frontend/app/src/App.tsx
@@ -1551,7 +1551,6 @@ export class App extends PureComponent<Props, State> {
       scriptRunId,
       scriptRunState,
       userSettings,
-      gitInfo,
       hideTopBar,
       hideSidebarNav,
       currentPageScriptHash,
@@ -1661,16 +1660,6 @@ export class App extends PureComponent<Props, State> {
                   developmentMode={developmentMode}
                   sendMessageToHost={
                     this.hostCommunicationMgr.sendMessageToHost
-                  }
-                  gitInfo={gitInfo}
-                  showDeployError={this.showDeployError}
-                  closeDialog={this.closeDialog}
-                  isDeployErrorModalOpen={
-                    this.state.dialog?.type === DialogType.DEPLOY_ERROR
-                  }
-                  loadGitInfo={this.sendLoadGitInfoBackMsg}
-                  canDeploy={
-                    this.sessionInfo.isSet && !this.sessionInfo.isHello
                   }
                   menuItems={menuItems}
                   metricsMgr={this.metricsMgr}
