--- frontend/app/src/components/StreamlitDialog/DeployErrorDialogs/RepoIsAhead.test.tsx
@@ -1,25 +1,0 @@
-/**
- * Copyright (c) Streamlit Inc. (2018-2022) Snowflake Inc. (2022)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import RepoIsAhead from "./RepoIsAhead"
-
-describe("DeployErrorDialog - RepoIsAhead", () => {
-  it("should render without crashing", () => {
-    const dialog = RepoIsAhead()
-
-    expect(dialog).toMatchSnapshot()
-  })
-})
