+++ frontend/app/src/components/StreamlitDialog/DeployErrorDialogs/index.ts
@@ -14,9 +14,6 @@
  * limitations under the License.
  */
 
-export { default as RepoIsAhead } from "./RepoIsAhead"
 export { default as DetachedHead } from "./DetachedHead"
-export { default as UntrackedFiles } from "./UntrackedFiles"
 export { default as ModuleIsNotAdded } from "./ModuleIsNotAdded"
-export { default as UncommittedChanges } from "./UncommittedChanges"
 export { default as NoRepositoryDetected } from "./NoRepositoryDetected"
